+++ test/helper.rb
@@ -9,3 +9,23 @@ MiniTest::Reporters.use! MiniTest::Reporters::SpecReporter.new
 vimrc = File.expand_path("../vimrc", __FILE__)
 $vim = Vimrunner::Server.new(:vimrc => vimrc).start
 Minitest::Unit.after_tests { $vim.kill }
+
+module WithTemporaryDirectory
+  def setup
+    super
+    # Mac has the temporary directory symlinked, so need File.realpath to
+    # match the paths that Vim returns.
+    @dir = File.realpath(Dir.mktmpdir) 
+  end
+
+  def teardown
+    FileUtils.remove_entry_secure @dir 
+    super
+  end
+end
+
+def touch(path, contents = nil)
+  FileUtils.mkdir_p File.dirname(path)
+  return FileUtils.touch(path) if contents.nil? || contents.empty?
+  File.open(path, "w") {|f| f.write contents }
+end