+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Unreleased
 - Adds `&include` pattern so Vim can recognize included/required files, e.g. for looking up keywords with `[I`.
+- Cleans `&path` from `/usr/include` for JavaScript files.
+- Adds a new superb `gf` handler to handle all relative and module paths, incl. support for `require(".")` to open `./index.js`. This is spot on how Node.js finds your requires.
+- Adds `<Plug>NodeGotoFile` should you want to remap Node.vim's file opener.
+- Opens files before directories should both, e.g. `./foo.js` and `./foo` exist. This matches Node.js's behavior.
 
 ## 0.2.0 (Jul 28, 2013)
 - Adds full support for navigating to module files by using `gf` on `require("any-module")`.